pyproject.toml
[tool.poetry]
name = "sql_scribe"
version = "0.1.0"
description = ""
authors = ["Tom Fynes"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.9.1"
XlsxWriter = "1.4.5"
colorama = "^0.4.4"
pyyaml = "5.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
